fix(security): fail-fast VehicleTelemetry guard and Long-overflow hardening
F-001 (VehicleTelemetry): add `require(raw.size >= 2)` constructor guard to
the JVM and iOS actual value classes (16-bit layout, bits 0-15). Truncated
buffers fail fast with IllegalArgumentException at construction instead of
a delayed, index-leaking AIOOBE at field access. The `expect` carries a
documenting comment only (expect cannot host an init body). Raw Int/Boolean
readBits getters are unchanged — the fast path stays caller-verified per
Tickets 06/08; checked accessors remain the never-throw path.

F-002 (KompactRuntime): promote fits()/readFloat()/readDouble() bounds
checks to Long arithmetic. Int `raw.size * 8` overflowed at 256 MiB,
silently classifying in-bounds reads as BoundsError.

F-003 (KompactFraming): nestedRegionOrNull bounds the region via Long
fits() and rejects `byteCount > Int.MAX_VALUE/8` before the `byteCount*8`
Int multiply, so a 0x7FFFFFFF length prefix fail-fasts to null
(TruncatedNested) instead of wrapping to a negative bit-length
(Pair(32,-8)).

TDD: 5 regression tests (constructor guard, 256 MiB boundary, 0x7FFFFFFF
wrap) red-then-green. Full :kompact:jvmTest BUILD SUCCESSFUL (rerun-tasks);
:kompact:apiCheck golden unchanged. Standards+Spec (RulingWasp) and
Security (ThickGopher) reviews both pass — no in-scope findings.

+++ kompact/src/commonMain/kotlin/ch/trancee/kompact/runtime/KompactRuntime.kt
@@ -125,7 +125,7 @@ public object KompactRuntime {
 
     /** Bounds-check: true iff [bitOffset]+[bitWidth] fits in [raw]. */
     public inline fun fits(raw: ByteArray, bitOffset: Int, bitWidth: Int): Boolean =
-        bitOffset >= 0 && bitWidth >= 1 && bitOffset + bitWidth <= raw.size * 8
+        bitOffset >= 0 && bitWidth >= 1 && bitOffset.toLong() + bitWidth.toLong() <= raw.size.toLong() * 8L
 
     /** Reads 1 bit at [bitOffset] as a checked [BooleanResult]. */
     public inline fun readBool(raw: ByteArray, bitOffset: Int): BooleanResult {
@@ -228,7 +228,7 @@ public object KompactRuntime {
 
     /** Reads 32 bits at [bitOffset] as a checked [FloatResult]. NaN is canonicalized (Ticket 04). */
     public inline fun readFloat(raw: ByteArray, bitOffset: Int): FloatResult {
-        if (bitOffset < 0 || bitOffset + 32 > raw.size * 8) {
+        if (bitOffset < 0 || bitOffset.toLong() + 32L > raw.size.toLong() * 8L) {
             return FloatResult.failure(KompactDecodeError.BoundsError)
         }
         val bits = readBitsLong(raw, bitOffset, 32).toInt()
@@ -237,7 +237,7 @@ public object KompactRuntime {
 
     /** Reads 64 bits at [bitOffset] as a checked [DoubleResult]. NaN is canonicalized (Ticket 04). */
     public inline fun readDouble(raw: ByteArray, bitOffset: Int): DoubleResult {
-        if (bitOffset < 0 || bitOffset + 64 > raw.size * 8) {
+        if (bitOffset < 0 || bitOffset.toLong() + 64L > raw.size.toLong() * 8L) {
             return DoubleResult.failure(KompactDecodeError.BoundsError)
         }
         val bits = readBitsLong(raw, bitOffset, 64)
